+++ packages/core/src/odata-common/filter/filterable.ts
@@ -6,6 +6,7 @@ import { FilterLink } from './filter-link';
 import { FilterList } from './filter-list';
 import { FilterLambdaExpression } from './filter-lambda-expression';
 import { BooleanFilterFunction } from './boolean-filter-function';
+import { UnaryFilter } from './unary-filter';
 
 /**
  * A union of all types that can be used for filtering.
@@ -17,6 +18,7 @@ export type Filterable<EntityT extends Entity> =
   | FilterLink<EntityT>
   | FilterList<EntityT>
   | FilterLambdaExpression<EntityT>
+  | UnaryFilter<EntityT>
   | BooleanFilterFunction<EntityT>;
 
 /**
@@ -92,7 +94,8 @@ export function toFilterableList<
   return filters.map(f => (f instanceof OneToManyLink ? f._filters : f));
 }
 
-// TODO:
-// Export function not<EntityT extends Entity>(expression: Filterable<EntityT>): Filterable<EntityT> {
-//   Return new FilterList([], expressions);
-// }
+export function not<EntityT extends Entity>(
+  filter: Filterable<EntityT>
+): UnaryFilter<EntityT> {
+  return new UnaryFilter(filter, 'not');
+}